implement truly observable for presetState collection, this way if state or collection of states is modified we update all state views in the editor
implement reassociation logic in a state between some ID and a node model to replace it with it. The data from the old ID is associated with the new node model, raises events when modifiy the state

fixed trulyobservable collection, not sure why this line was commented out... no note.

presetStateViewModel generates buttons for each serialized node contained in the state model

it also handles button presses and calls the correct methods on different states to reassociate states if needed

PresetModelView recreates all statesViews and thus all buttons when the workspace nodes in graph are modified or if states are modified, for instance if nodes are deleted from the graph we might need to turn some buttons red.

removed double scroll view in xaml